ref(installer): LUKS key files move to disk/keyfiles.py as KeyFileGenerator
* generate() returns the in-target paths mkinitcpio FILES embeds
* keyfile dir shared with cryptenroll through KEYFILE_DIR
